Move ComboBox to components-styled folder
