+++ packages/app/src/components-styled/combo-box/combo-box.scss
@@ -1,5 +1,9 @@
 @import '~/scss/variables.scss';
 
+/**
+ * These are global styles, not styles within the component itself.
+ */
+
 [data-reach-combobox] {
   position: relative;
