+++ packages/app/src/domain/layout/components/municipality-combo-box.tsx
@@ -1,5 +1,5 @@
 import { useRouter } from 'next/router';
-import { ComboBox } from '~/components/comboBox';
+import { ComboBox } from '~/components-styled/combo-box/combo-box';
 import municipalities from '~/data/municipalSearchData';
 import siteText from '~/locale/index';
 import { useBreakpoints } from '~/utils/useBreakpoints';
