+++ packages/app/src/domain/layout/components/safety-region-combo-box.tsx
@@ -1,5 +1,5 @@
 import { useRouter } from 'next/router';
-import { ComboBox } from '~/components/comboBox';
+import { ComboBox } from '~/components-styled/combo-box/combo-box';
 import safetyRegions from '~/data/index';
 import siteText from '~/locale/index';
 import { useBreakpoints } from '~/utils/useBreakpoints';
